# backend/tests/openapi_snapshot.json
{
  "components": {
    "schemas": {
      "AggregationLabelCatalogOut": {
        "properties": {
          "history_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "History Error"
          },
          "history_series_count": {
            "title": "History Series Count",
            "type": "integer"
          },
          "history_status": {
            "enum": [
              "ok",
              "unconfigured",
              "error"
            ],
            "title": "History Status",
            "type": "string"
          },
          "labels": {
            "items": {
              "$ref": "#/components/schemas/AggregationLabelOut"
            },
            "title": "Labels",
            "type": "array"
          },
          "lookback_hours": {
            "enum": [
              24,
              168,
              720
            ],
            "title": "Lookback Hours",
            "type": "integer"
          }
        },
        "required": [
          "lookback_hours",
          "history_status",
          "history_error",
          "history_series_count",
          "labels"
        ],
        "title": "AggregationLabelCatalogOut",
        "type": "object"
      },
...
      "AggregationRuleOut": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "group_by_labels": {
            "items": {
              "type": "string"
            },
            "title": "Group By Labels",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "matchers": {
            "items": {
              "$ref": "#/components/schemas/AggregationMatcherIn"
            },
            "title": "Matchers",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "priority": {
            "default": 100,
            "title": "Priority",
            "type": "integer"
          },
          "source_scope": {
            "$ref": "#/components/schemas/SourceScope"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "id",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "AggregationRuleOut",
        "type": "object"
      },
      "AggregationRulePreviewOut": {
        "properties": {
          "by_source": {
            "items": {
              "$ref": "#/components/schemas/AggregationPreviewSourceOut"
            },
            "title": "By Source",
            "type": "array"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/AggregationPreviewGroupOut"
            },
            "title": "Groups",
            "type": "array"
          },
          "matcher_alert_count": {
            "title": "Matcher Alert Count",
            "type": "integer"
          },
          "proposed_group_count": {
            "title": "Proposed Group Count",
            "type": "integer"
          },
          "selected_alert_count": {
            "title": "Selected Alert Count",
            "type": "integer"
          }
        },
        "required": [
          "matcher_alert_count",
          "selected_alert_count",
          "proposed_group_count",
          "groups"
        ],
        "title": "AggregationRulePreviewOut",
        "type": "object"
      },
...
      "AggregationRulePut": {
        "properties": {
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "group_by_labels": {
            "items": {
              "type": "string"
            },
            "title": "Group By Labels",
            "type": "array"
          },
          "matchers": {
            "items": {
              "$ref": "#/components/schemas/AggregationMatcherIn"
            },
            "title": "Matchers",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "priority": {
            "default": 100,
            "title": "Priority",
            "type": "integer"
          },
          "source_scope": {
            "$ref": "#/components/schemas/SourceScope"
          }
        },
        "required": [
          "name"
        ],
        "title": "AggregationRulePut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Alert Workbench",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/aggregation-labels": {
      "get": {
        "operationId": "get_aggregation_labels_api_aggregation_labels_get",
        "parameters": [
          {
            "in": "query",
            "name": "lookback_hours",
            "required": false,
            "schema": {
              "default": 168,
              "title": "Lookback Hours",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregationLabelCatalogOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Aggregation Labels"
      }
    },
    "/api/aggregation-rules": {
      "get": {
        "operationId": "get_aggregation_rules_api_aggregation_rules_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AggregationRuleOut"
                  },
                  "title": "Response Get Aggregation Rules Api Aggregation Rules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Aggregation Rules"
      },
      "post": {
        "operationId": "create_rule_api_aggregation_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AggregationRulePut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregationRuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Rule"
      }
    },
    "/api/aggregation-rules/preview": {
      "post": {
        "operationId": "preview_rule_api_aggregation_rules_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AggregationRulePreviewPut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregationRulePreviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Rule"
      }
    },
    "/api/aggregation-rules/{rule_id}": {
      "put": {
        "operationId": "update_rule_api_aggregation_rules__rule_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AggregationRulePut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregationRuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Rule"
      }
    },
...
    "/api/incidents": {
      "get": {
        "operationId": "list_incidents_api_incidents_get",
        "parameters": [
          {
            "description": "Filter by source_state",
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source_state",
              "title": "State"
            }
          },
          {
            "description": "Filter by severity",
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by severity",
              "title": "Severity"
            }
          },
          {
            "description": "Comma separated or repeated source IDs",
            "in": "query",
            "name": "source_ids",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma separated or repeated source IDs",
              "title": "Source Ids"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IncidentSummary"
                  },
                  "title": "Response List Incidents Api Incidents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Incidents"
      }
    },
...
  }
}